Added support for pseudo-instructions.
These take the format instruction!(operand). The intent is
to provide support for common operations that don't neccessarily
correspond to a specific assembler mnemonic, or wouldn't be
found in the ARM technical reference.

The first supported pseudo is `thumb-addr!(&ref)`, which expects a
label reference and relaces it with <label-address> | 1.

An obvious objection to this pseudo is that it's too specific. A more
general pseudo like `math!(&ref | 1)` would be applicable to a variety
of usecases. I agree, but a) I wanted something I could use immediately,
and b) I didn't want to get tangled up in a more general-purpose
implementation when it wasn't strictly neccessary.

It's possible that thumb-addr! gets removed in lieu of something more
general, but this is fine for now.